refactor: modularize Ethereum validators
