+++ frontend/app/src/components/accounts/EthStakingValidators.vue
@@ -1,194 +1,57 @@
 <script setup lang="ts">
-import type { ContextColorsType, DataTableColumn } from '@rotki/ui-library';
 import type { StakingValidatorManage } from '@/composables/accounts/blockchain/use-account-manage';
-import type { EthereumValidator, EthereumValidatorRequestPayload } from '@/types/blockchain/accounts';
-import { Blockchain, One, Zero } from '@rotki/common';
+import type { EthereumValidator } from '@/types/blockchain/accounts';
 import Eth2ValidatorLimitRow from '@/components/accounts/blockchain/eth2/Eth2ValidatorLimitRow.vue';
+import ValidatorStatus from '@/components/accounts/staking/eth/ValidatorStatus.vue';
 import AmountDisplay from '@/components/display/amount/AmountDisplay.vue';
 import PercentageDisplay from '@/components/display/PercentageDisplay.vue';
 import RowActions from '@/components/helper/RowActions.vue';
 import RowAppend from '@/components/helper/RowAppend.vue';
 import TableFilter from '@/components/table-filter/TableFilter.vue';
-import { useAccountDelete } from '@/composables/accounts/blockchain/use-account-delete';
-import { useEthStaking } from '@/composables/blockchain/accounts/staking';
-import { type Filters, type Matcher, useEthValidatorAccountFilter } from '@/composables/filters/eth-validator';
-import { usePaginationFilters } from '@/composables/use-pagination-filter';
-import { useBlockchainBalances } from '@/modules/balances/use-blockchain-balances';
+import { useEthValidatorData } from '@/composables/staking/eth/use-eth-validator-data';
+import { useEthValidatorOperations } from '@/composables/staking/eth/use-eth-validator-operations';
+import { useEthValidatorUtils } from '@/composables/staking/eth/use-eth-validator-utils';
 import HashLink from '@/modules/common/links/HashLink.vue';
-import { usePriceUtils } from '@/modules/prices/use-price-utils';
-import { TableId, useRememberTableSorting } from '@/modules/table/use-remember-table-sorting';
-import { useBlockchainValidatorsStore } from '@/store/blockchain/validators';
 import { useGeneralSettingsStore } from '@/store/settings/general';
-import { useStatusStore } from '@/store/status';
-import { useTaskStore } from '@/store/tasks';
 import { SavedFilterLocation } from '@/types/filtering';
-import { Section } from '@/types/status';
-import { TaskType } from '@/types/task-type';
 
 const emit = defineEmits<{
-  (e: 'edit', value: StakingValidatorManage): void;
+  edit: [value: StakingValidatorManage];
 }>();
 
 const { t } = useI18n({ useScope: 'global' });
 
-const selected = ref<number[]>([]);
-
-const blockchainValidatorsStore = useBlockchainValidatorsStore();
-const { fetchValidators } = blockchainValidatorsStore;
-const { ethStakingValidators } = storeToRefs(blockchainValidatorsStore);
-const { currencySymbol } = storeToRefs(useGeneralSettingsStore());
-const { showConfirmation } = useAccountDelete();
-const { fetchEthStakingValidators } = useEthStaking();
-const { useExchangeRate } = usePriceUtils();
-const { fetchBlockchainBalances } = useBlockchainBalances();
-
+// Use composables
 const {
-  fetchData,
+  cols,
+  ethStakingValidators,
   filters,
   matchers,
   pagination,
+  rows,
+  selected,
   sort,
-  state: rows,
-} = usePaginationFilters<
-  EthereumValidator,
-  EthereumValidatorRequestPayload,
-  Filters,
-  Matcher
->(fetchValidators, {
-  defaultSortBy: {
-    column: 'index',
-    direction: 'desc',
-  },
-  filterSchema: () => useEthValidatorAccountFilter(t),
-  history: 'router',
-});
+} = useEthValidatorData();
 
-const cols = computed<DataTableColumn<EthereumValidator>[]>(() => {
-  const currency = { symbol: get(currencySymbol) };
-  return [
-    {
-      cellClass: 'py-0',
-      key: 'index',
-      label: t('common.validator_index'),
-      sortable: true,
-    },
-    {
-      cellClass: 'py-0',
-      key: 'publicKey',
-      label: t('eth2_input.public_key'),
-      sortable: true,
-    },
-    {
-      cellClass: 'py-0',
-      key: 'status',
-      label: t('common.status'),
-      sortable: true,
-    },
-    {
-      align: 'end',
-      cellClass: 'py-0',
-      key: 'amount',
-      label: t('common.amount'),
-      sortable: true,
-    },
-    {
-      align: 'end',
-      cellClass: 'py-0',
-      key: 'usdValue',
-      label: t('account_balances.headers.usd_value', currency),
-      sortable: true,
-    },
-    {
-      align: 'end',
-      cellClass: 'py-0',
-      key: 'ownershipPercentage',
-      label: t('common.ownership'),
-      sortable: false,
-    },
-    {
-      align: 'end',
-      cellClass: '!p-0',
-      key: 'actions',
-      label: t('common.actions_text'),
-    },
-  ];
-});
+const {
+  accountOperation,
+  confirmDelete,
+  deleteSelected,
+  edit: editValidator,
+  refresh,
+} = useEthValidatorOperations();
 
-useRememberTableSorting<EthereumValidator>(TableId.ETH_STAKING_VALIDATORS, sort, cols);
-
-const { useIsTaskRunning } = useTaskStore();
-const { isLoading } = useStatusStore();
-
-const loading = isLoading(Section.BLOCKCHAIN, Blockchain.ETH2);
-
-const accountOperation = logicOr(
-  useIsTaskRunning(TaskType.ADD_ACCOUNT),
-  useIsTaskRunning(TaskType.REMOVE_ACCOUNT),
-  loading,
-);
+const { currencySymbol } = storeToRefs(useGeneralSettingsStore());
+const { getOwnershipPercentage, useTotal } = useEthValidatorUtils();
+const total = useTotal(rows);
 
 function edit(account: EthereumValidator) {
-  const { index, ownershipPercentage, publicKey } = account;
-  const state: StakingValidatorManage = {
-    chain: Blockchain.ETH2,
-    data: {
-      ownershipPercentage: ownershipPercentage ?? '100',
-      publicKey,
-      validatorIndex: index.toString(),
-    },
-    mode: 'edit',
-    type: 'validator',
-  };
-  emit('edit', state);
-}
-
-const colorMap: Record<string, ContextColorsType | undefined> = {
-  active: 'success',
-  consolidated: 'secondary',
-  exited: 'error',
-  exiting: 'warning',
-  pending: 'info',
-};
-
-function getColor(status: string): ContextColorsType | undefined {
-  return colorMap[status] ?? undefined;
-}
-
-function getOwnershipPercentage(row: EthereumValidator): string {
-  return row.ownershipPercentage || '100';
-}
-
-async function refresh() {
-  await fetchEthStakingValidators();
-  await fetchBlockchainBalances({
-    blockchain: Blockchain.ETH2,
-    ignoreCache: true,
-  });
-}
-
-function confirmDelete(item: EthereumValidator) {
-  showConfirmation({
-    data: [item],
-    type: 'validator',
-  });
+  emit('edit', editValidator(account));
 }
 
 function deleteSelectedValidators() {
-  const items = get(rows).data.filter(item => get(selected).includes(item.index));
-  showConfirmation({
-    data: items,
-    type: 'validator',
-  });
+  deleteSelected(get(rows).data, get(selected));
 }
-
-const total = computed(() => {
-  const mainCurrency = get(currencySymbol);
-  return (get(rows).totalUsdValue || Zero).multipliedBy(get(useExchangeRate(mainCurrency)) ?? One);
-});
-
-watchImmediate(ethStakingValidators, async () => {
-  await fetchData();
-});
 
 defineExpose({
   refresh,
@@ -278,34 +141,7 @@ defineExpose({
         />
       </template>
       <template #item.status="{ row }">
-        <RuiChip
-          size="sm"
-          :color="getColor(row.status)"
-          content-class="text-xs inline-flex gap-1 items-center"
-          class="uppercase font-semibold"
-        >
-          {{ row.status }}
-
-          <template v-if="row.consolidatedInto">
-            <RuiTooltip
-              persist-on-tooltip-hover
-              :open-delay="300"
-            >
-              <template #activator>
-                <RuiIcon
-                  name="lu-merge"
-                  size="16"
-                  class="bg-rui-grey-200 text-rui-grey-800 rounded-full p-0.5 -mr-1.5 cursor-pointer"
-                />
-              </template>
-              <div>{{ t('blockchain_balances.validators.consolidated') }}</div>
-              <HashLink
-                location="eth2"
-                :text="row.consolidatedInto.toString()"
-              />
-            </RuiTooltip>
-          </template>
-        </RuiChip>
+        <ValidatorStatus :validator="row" />
       </template>
       <template #item.amount="{ row }">
         <AmountDisplay
